pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toolipie"
version = "0.1.0"
description = "Toolipie — personal CLI toolbox (Markdown to DOCX/PDF via Pandoc)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12,<1",
  "rich>=13,<14",
  "PyYAML>=6,<7",
  "python-dotenv>=1,<2",
  "pypandoc>=1,<2",
  "markdown-it-py>=3,<5",
  "weasyprint>=61,<63",
  "Pillow>=10,<11",
  "pypdfium2>=4,<5",
  "opencv-python-headless>=4.10,<5",
]

[project.scripts]
toolipie = "toolipie.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["toolipie*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
